TrackMate_: rework the DownSampled segmenter for better efficiency
Use a simpler downsmapling scheme.
Do segmentation directly with correct scale, which dismisses the need to
scale result up afterwards.
It works suprisingly well.

+++ src-plugins/TrackMate_/fiji/plugin/trackmate/segmentation/DownSampleLogSegmenterConfigurationPanel.java
@@ -48,7 +48,7 @@ public class DownSampleLogSegmenterConfigurationPanel extends LogSegmenterConfig
 	public SegmenterSettings getSegmenterSettings() {
 		LogSegmenterSettings lss = (LogSegmenterSettings) super.getSegmenterSettings();
 		DownSampleLogSegmenterSettings dss = copyToDSLSS(lss);
-		dss.downSamplingFactor = Float.parseFloat(jTextFieldDownSample.getText());
+		dss.downSamplingFactor = Integer.parseInt(jTextFieldDownSample.getText());
 		return dss;
 	}
